função para deletar produto do carrinho, e mudando estrutura da função getProducts

+++ src/screens/CartScreen/index.js
@@ -1,9 +1,9 @@
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import ProductCart from '../../components/ProductCart';
 import Api from '../../Api';
 import auth from '@react-native-firebase/auth';
-import {useNavigation} from '@react-navigation/native';
-import {useSelector} from 'react-redux';
+import { useNavigation } from '@react-navigation/native';
+import { useSelector } from 'react-redux';
 import EmptyCart from '../../assets/svg/empty_cart.svg';
 import LoginSvg from '../../assets/svg/login.svg';
 import LoadingScreen from '../../components/LoadingComponent';
@@ -11,7 +11,7 @@ import LoadingScreen from '../../components/LoadingComponent';
 
 import {
     Container,
-    
+
     FlatView,
     Flat,
 
@@ -29,34 +29,37 @@ export default () => {
     const [userInfo, setUserInfo] = useState([]);
     const [arrayCart, setArrayCart] = useState([]);
     const [loading, setLoading] = useState(true);
+    const [id, setId] = useState();
 
-    const userLogin = useSelector(state=>state.user.email);
+    const userLogin = useSelector(state => state.user.email);
 
     const navigation = useNavigation();
 
     var subtotal = 0;       // Variável que pega o valor total dos produtos que estão no carrinho
     var quantidadeTotal = 0;
 
-    if(arrayCart) {
+    if (arrayCart) {
         arrayCart.forEach(item => {     // Função foreach que pega o subtotal
             subtotal += item.price * item.quantidade
             quantidadeTotal += item.quantidade;
         })
     }
-    
-    
 
-    if(userLogin) {
+
+
+    if (userLogin) {
         const userId = auth().currentUser.uid;
 
         useEffect(() => {
             const getProducts = () => {
                 setArrayCart([]);
-                
+
                 let json = Api.getProductsCart(userId, setArrayCart);
                 setArrayCart(json)
+                setId(userId);
+
             }
-    
+
             getProducts();
         }, [])
 
@@ -79,7 +82,7 @@ export default () => {
 
 
     const NoProduct = () => {
-        return(
+        return (
             <NoInfoView>
                 <EmptyCart width="200px" height="200px" />
                 <NoInfoText>O seu carrinho está vazio </NoInfoText>
@@ -88,7 +91,7 @@ export default () => {
     }
 
     const NoUserLogin = () => {
-        return(
+        return (
             <NoInfoView>
                 <LoginSvg width="200px" height="200px" />
                 <NoInfoText>Para ver seu carrinho, por favor faça o Login</NoInfoText>
@@ -107,21 +110,25 @@ export default () => {
             userAdress = item.adress;
         })
 
-        let res = await Api.setUserOrder(userId, userAdress, userName, arrayCart, subtotal, quantidadeTotal, navigation);
+        if (userAdress.length > 0) {
+            Api.setUserOrder(userId, userAdress, userName, arrayCart, subtotal, quantidadeTotal, navigation);
 
-        let del = await Api.deleteCart(userId);
+            Api.deleteCart(userId);
+        } else {
+            alert('Escolha um endereço para a entraga antes...')
+        }
     }
 
-    return(
+    return (
         <Container>
             {/* Quando abre a tela vai ferificar o tempo de loading que é 2s */}
-            {!loading ? 
+            {!loading ?
                 <>
                     {/* Depois verifica se tem algum usuário logado */}
-                    {userLogin ? 
+                    {userLogin ?
                         <>
                             {/* Depois verifica se o usuário tem algum produto no carrinho */}
-                            {arrayCart ? 
+                            {arrayCart ?
                                 <>
                                     <FlatView>
                                         <Flat
@@ -133,31 +140,31 @@ export default () => {
                                                 </>
                                             }
                                             data={arrayCart}
-                                            renderItem={({item}) => <ProductCart data={item} />}
+                                            renderItem={({ item }) => <ProductCart data={item} userId={id} />}
                                             keyExtractor={(item) => item.id}
-                                        /> 
-                                    </FlatView> 
-                                    
-                                    <DefaultBtn onPress={setUserOrder}>
+                                        />
+                                    </FlatView>
+
+                                    <DefaultBtn bgColor={arrayCart && subtotal && quantidadeTotal ? '#ea1d2c' : '#aaa'} disabled={arrayCart && subtotal && quantidadeTotal ? false : true} onPress={setUserOrder}>
                                         <DefaultBtnText>Finalizar Pedido</DefaultBtnText>
-                                    </DefaultBtn> 
+                                    </DefaultBtn>
                                 </>
-                            :
+                                :
                                 // Se não tiver produto no carrinho, retorna a função que retorna um componente NoProduct
-                                <NoProduct />   
-                            }        
+                                <NoProduct />
+                            }
                         </>
 
-                    :
+                        :
                         // Se não tiver usuário logado, retorna a função que retorna um componente NoUserLogin
                         <NoUserLogin />
                     }
                 </>
 
-            : 
+                :
                 // Durante os 2s de loading mostra a função de LoadingScreen que é apenas um indicador de carregamento
                 <LoadingScreen />
-            }        
+            }
         </Container>
     );
 }
